[LayoutNG] Fix shape exclusion tracking.
The DerivedGeometry class gets moved from exclusion-space to
exclusion-space as a fast way of determining layout opportunities (where
lines, etc get placed).

For speed we only allocate "NGShapeExclusions" if we've seen any shapes
within the formatting context.

This crash was due to an inconsistent state of the:
NGExclusionSpaceInternal::track_shape_exclusions_ flag.

"other" NGExclusionSpaceInternal
 - derived_geometry_ = (non-null)
 - track_shape_exclusions_ = false
 - num_exclusions_ = 1u

"this" NGExclusionSpaceInternal
 - derived_geometry_ = null
 - track_shape_exclusions_ = true
 - num_exclusions_ = 2u (2nd exclusion has shape data).

Within the NGExclusionSpaceInternal::MoveAndUpdateDerivedGeometry method
we moved the derived geometry over to the "this" exclusion space.

However we didn't also move the track_shape_exclusions_ flag as a result
we'd try to add the 2nd shape exclusion, and a NPE was hit.

Bug: 962135, 962272, 962052
Change-Id: Id095bd2924f83ba6b2dbe90446d1f4613de09933