Allow empty backdrop_filter_bounds
Previous to this CL, an empty rect was used to signal that no clipping
should be performed on the backdrop filtered image. However, in the
corner case that the element with backdrop-filter was zero size, this
would cause no clipping to be performed, which is incorrect. With this
CL, backdrop_filter_bounds is base::Optional. There is also a test
for this case.

Change-Id: I8d3d1e383ccf807f1e4f569847bd828a8d04fd3d